Fix int() label cast in VideoGroupDataset + update preprint/rebuttal docs
- video_group_dataset.py: Change label = int(row["label"]) to float() to
  stop truncating regression labels. The int() cast quantized z-scored RVSP
  labels to ~5 bins, with 78% mapping to 0 (the mean). All ICML preprint
  RVSP numbers (Table 4) were artifacts of this quantization.

- experiment-issues.md: Add forensic analysis of ICML RVSP numbers showing
  all models predicted ~35 mmHg constantly (prediction std=0.03). Document
  EchoMAE 170x-low LR with inverted cosine schedule. Update impact table
  to mark preprint RVSP as invalid.

- 08-rebuttal-v2.md: Expand controlled comparison from 1 pair to 3-way
  epoch-matched study (JEPA-L vs BYOL-L vs MAE-L, all 50ep on MIMIC 525K).
  Add corrected EchoMAE retrain plan with standard LR. Update training
  status and BYOL learning curve.

+++ src/datasets/video_group_dataset.py
@@ -278,7 +278,7 @@ class VideoGroupDataset(Dataset):
             index = np.random.randint(len(self))
 
     def _get_item_row(self, row):
-        label = int(row["label"])
+        label = float(row["label"])
     
         # ---- collect URIs and initial presence flags from CSV ----
         uris, present = [], []
